cleanup wrapmode, fix bonus interval
when consuming an item, set the time of next bonus then (instead of just
when spawning). at spawn time we also factor in the time a bonus might
be alive (can be shortened by eating it).

bonus actions (i.e. wraparound) is now a bitmap. thus, more than 1
action might be active at the same time.